Extract change recording utils to a separate file
